added new component Profile; corrected App.tsx render code; corrected styles in Menu and Item components
