+++ src/components/Menu/Item/Item.tsx
@@ -1,17 +1,16 @@
 import React from 'react';
 import { FC } from 'react';
+import { cnMenu } from '../Menu.classname';
 
 type ItemProps = {
-  listItemClass: string;
   text: string;
   listStyleImage: {};
-  spanListItemClass: string;
 }
 
-const Item: FC<ItemProps> = ({ listItemClass, text, listStyleImage, spanListItemClass }) => {
+const Item: FC<ItemProps> = ({ text, listStyleImage, }) => {
   return (
-    <li className={listItemClass} style={listStyleImage}>
-      <span className={spanListItemClass}>{text}</span>
+    <li className={cnMenu('ListItem')} style={listStyleImage}>
+      <span className={cnMenu('ListItemSpan')}>{text}</span>
     </li>
   );
 }
